+++ bloom/ws/src/server.rs
@@ -333,7 +333,8 @@ async fn process_messages<C>(
                         let _ = sink.lock().await.send(Message::Pong(payload)).await;
                     }
                     Some(Ok(_)) => {
-                        // 非テキスト（バイナリ等）はメディア中継しない方針: 1003 Closeを返し、room_idをクリアして状態変化を避ける
+                        // 非テキスト（バイナリ等）はシグナリング外=メディア誤送信扱い。
+                        // leave_roomを発火させないためroom_idをクリアし、1003 Closeで拒否して中継しない。
                         handler.room_id = None;
                         let _ = sink
                             .lock()
