+++ bloom/ws/tests/common.rs
@@ -1,8 +1,9 @@
 use bloom_api::ServerToClient;
 use bloom_ws::{start_ws_server, MockCore, SharedCore, WsServerHandle};
-use futures_util::StreamExt;
+use futures_util::{SinkExt, StreamExt};
+use std::sync::{Arc, Mutex};
 use tokio::net::TcpStream;
-use tokio_tungstenite::{MaybeTlsStream, WebSocketStream};
+use tokio_tungstenite::{connect_async, MaybeTlsStream, WebSocketStream};
 use tokio_tungstenite::tungstenite::protocol::Message;
 
 pub async fn spawn_bloom_ws_server_with_core(
@@ -30,6 +31,50 @@ pub async fn recv_server_msg(
             panic!("ws closed before receiving message");
         }
     }
+}
+
+/// CreateRoomするクライアントAとJoinするクライアントBを起動し、ID類を返すヘルパー。
+#[allow(dead_code)]
+pub async fn setup_room_with_two_clients(
+    server_url: &str,
+    core_arc: &Arc<Mutex<MockCore>>,
+) -> (
+    WebSocketStream<MaybeTlsStream<TcpStream>>, // ws_a
+    WebSocketStream<MaybeTlsStream<TcpStream>>, // ws_b
+    String,                                     // room_id
+    String,                                     // a_id
+    String,                                     // b_id
+) {
+    let (mut ws_a, _) = connect_async(server_url).await.expect("connect A");
+    ws_a.send(Message::Text(r#"{"type":"CreateRoom"}"#.into()))
+        .await
+        .expect("send create room");
+    let room_created = recv_server_msg(&mut ws_a).await;
+    let (room_id, a_id) = match room_created {
+        ServerToClient::RoomCreated { room_id, self_id } => (room_id, self_id),
+        other => panic!("expected RoomCreated, got {:?}", other),
+    };
+
+    let (mut ws_b, _) = connect_async(server_url).await.expect("connect B");
+    ws_b.send(Message::Text(format!(
+        r#"{{"type":"JoinRoom","room_id":"{room_id}"}}"#,
+        room_id = room_id
+    )))
+    .await
+    .expect("send join room");
+    // join完了を待つ
+    let _ = recv_server_msg(&mut ws_b).await;
+
+    // Join時に付与されたparticipant_idをCore側の記録から取得
+    let b_id = {
+        let core = core_arc.lock().expect("lock core");
+        core.join_room_calls
+            .last()
+            .map(|(_, p)| p.to_string())
+            .expect("b id recorded")
+    };
+
+    (ws_a, ws_b, room_id, a_id, b_id)
 }
 
 // minimal helpers shared across test files
